requirements.txt
Flask==3.0.2
waitress==2.0.0
SQLAlchemy==2.0.29
pymysql==1.1.0
sqlalchemy-utils==0.41.2
PyJWT==2.8.0
pytest==8.1.1
firebase_admin==6.5.0